Inaccurate data
"Currently, an extract, transform, and load \(ETL\) job can access resources in a single subnet only\. If you have multiple data stores in a job, they must be on the same subnet\."

This is outdated and inaccurate:

"That is an old statement, It used to be that having resources in multiple subnets would confuse the job engine and it would not know which subnet to launch it's ENI in, therefore it would throw a Multi-subnet error and break. That was fixed a long time ago." 

I can confirm that accessing resources in multiple subnets from Glue is not an issue anymore, I do it frequently.